% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_closest_players.R
\name{get_closest_players}
\alias{get_closest_players}
\title{get_closest_players}
\usage{
get_closest_players(df)
}
\arguments{
\item{df}{a dataframe built with get_play_sf}
}
\value{
df with an extra column distance_all with each player's distance to each opposing player
}
\description{
get_closest_players
}
