% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_geometry.R
\name{get_convex_hull_def}
\alias{get_convex_hull_def}
\title{Compute convex hull for offensive players}
\usage{
get_convex_hull_def(df)
}
\arguments{
\item{df}{df with players_def column}
}
\value{
df with convex_hull_def column with convex hulls
}
\description{
Compute convex hull for offensive players
}
\examples{
get_convex_hull_def(df)
}
