% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_geometry.R
\name{get_convex_hull_off}
\alias{get_convex_hull_off}
\title{Compute convex hull for offensive players}
\usage{
get_convex_hull_off(df)
}
\arguments{
\item{df}{}
}
\value{
df with players_off column
}
\description{
Compute convex hull for offensive players
}
\examples{
get_convex_hull_off(df)
}
