% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_geometry.R
\name{get_delaunay_triangles_off}
\alias{get_delaunay_triangles_off}
\title{Compute Delaunay Triangles for offensive players}
\usage{
get_delaunay_triangles_off(df)
}
\arguments{
\item{df}{df with players_off column}
}
\value{
df with convex_hull_off column with convex hulls
}
\description{
Compute Delaunay Triangles for offensive players
}
\examples{
get_convex_hull_off(df)
}
