% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_play.R
\name{get_nested_sf}
\alias{get_nested_sf}
\title{get_nested_transformations
Gets all combinations of interest between objects}
\usage{
get_nested_sf(df)
}
\arguments{
\item{df}{with column nflId,playId and gameId, position_type_1
nflId=9999 if football}
}
\value{

}
\description{
get_nested_transformations
Gets all combinations of interest between objects
}
