% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_play.R
\name{get_play}
\alias{get_play}
\title{get_play
Searches for a play without any transformations}
\usage{
get_play(playid = NULL, gameid)
}
\arguments{
\item{playid}{if NULL (default), will retrieve the whole game}

\item{gameid}{}
}
\value{

}
\description{
get_play
Searches for a play without any transformations
}
\examples{
df=get_play(1946,2018101412)
}
