% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_functions.R
\name{get_play_boundary}
\alias{get_play_boundary}
\title{get_play_boundary}
\usage{
get_play_boundary(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
get_play_boundary
}
\examples{
get_play_boundary(sf_sample)
}
