% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_play.R
\name{get_play_sf}
\alias{get_play_sf}
\title{get_play_sf
Gets a play or a whole game,
transforming all coordintates to a simple feature list column, and embedding tracking data into every object in the field
After converting to sf, it will create all combinations of interest between the objects, like all, players, football, players off
and players def. Objects will be repeated and this might not be efficient if only using a certain combination.}
\usage{
get_play_sf(playid = NULL, gameid, ...)
}
\value{

}
\description{
get_play_sf
Gets a play or a whole game,
transforming all coordintates to a simple feature list column, and embedding tracking data into every object in the field
After converting to sf, it will create all combinations of interest between the objects, like all, players, football, players off
and players def. Objects will be repeated and this might not be efficient if only using a certain combination.
}
\examples{
df=get_play_sf(playid=NULL,gameid)
}
