% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_functions.R
\name{get_play_stats}
\alias{get_play_stats}
\title{get_play_stats
Applies a series of transformations to the each frame to tag important information}
\usage{
get_play_stats(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
get_play_stats
Applies a series of transformations to the each frame to tag important information
}
\examples{
get_play_stats(sf_sample)
}
