% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_functions.R
\name{get_sf_vornoi}
\alias{get_sf_vornoi}
\title{get_sf_vornoi}
\usage{
get_sf_vornoi(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
get_sf_vornoi
}
\examples{
get_play_boundary(sf_sample) \%>\% get_sf_vornoi()
}
