% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_functions.R
\name{get_vornoi_analytics}
\alias{get_vornoi_analytics}
\title{get_vornoi_analytics}
\usage{
get_vornoi_analytics(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
get_vornoi_analytics
}
\examples{
get_play_boundary(sf_sample) \%>\% get_sf_vornoi() \%>\% get_play_stats() \%>\% get_vornoi_analytics()
}
