% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_functions.R
\name{get_vornoi_transformations}
\alias{get_vornoi_transformations}
\title{get_vornoi_transformations
Apply all transformations needed to compute Vornoi areas and their areas.
For examples of how this can be used check out https://www.kaggle.com/est092/vornoi-areas}
\usage{
get_vornoi_transformations(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
get_vornoi_transformations
Apply all transformations needed to compute Vornoi areas and their areas.
For examples of how this can be used check out https://www.kaggle.com/est092/vornoi-areas
}
\examples{
get_vornoi_transformations(sf_sample)
}
