% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_db.R
\name{load_games}
\alias{load_games}
\title{load_games}
\usage{
load_games(src_dir = "data")
}
\arguments{
\item{src_dir}{specify the working directory of the necesary files
Point to directory with game index}
}
\value{
game index
}
\description{
load_games
}
\examples{
load_games("../roboShutdown/data")
}
