% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_db.R
\name{load_players}
\alias{load_players}
\title{load_players}
\usage{
load_players(src_dir = "data")
}
\arguments{
\item{src_dir}{specify the working directory of the necesary files}
}
\value{
players_index
}
\description{
load_players
}
\examples{
load_players("../roboShutdown/data")
}
