% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_db.R
\name{load_plays}
\alias{load_plays}
\title{load_plays}
\usage{
load_plays(src_dir = "data")
}
\arguments{
\item{src_dir}{specify the working directory of the necesary files
Point to directory with play index}
}
\value{
all connections loaded as lazy dataframes
}
\description{
load_plays
}
\examples{
load_plays("../roboShutdown/data")
}
