% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_db.R
\name{load_tracking_data}
\alias{load_tracking_data}
\title{load_tracking_data}
\usage{
load_tracking_data(src_dir = "data", week = 1, file_prefix = "week")
}
\arguments{
\item{src_dir}{specify the working directory of the necesary files}

\item{week}{week to be loaded, can be any week in the folder or all .}

\item{file_prefix}{default prefix to tracking data filenames, default is week like the original 2020 databowl files
Week name must look like this: week\link{0-9}+ , example:week1.csv
if you want to use more than week at a time, you should look into creating a database and using dbplyr
If you plan on using different play by play data, you should make sure it has the following columns: nflId, GameId,playId}
}
\value{
all connections loaded as lazy dataframes
}
\description{
load_tracking_data
}
\examples{
load_tracking_data()
}
