% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_plot.R
\name{plot_vornoi_analytics}
\alias{plot_vornoi_analytics}
\title{plot_vornoi_analytics}
\usage{
plot_vornoi_analytics(df, file_dir = "output/", type = "all")
}
\arguments{
\item{file_dir}{directory were file will be created, will create a dir for the game, and will try to put the plots there}

\item{type}{types of plots to make, will default to all, but you can pass a vector of: "players","stacked","team","all"}
}
\value{

}
\description{
plot_vornoi_analytics
}
