% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_plot.R
\name{plot_vornoi_animation}
\alias{plot_vornoi_animation}
\title{get_vornoi_animation
Will need to have gganimate installed}
\usage{
plot_vornoi_animation(df, file_dir = "output/")
}
\arguments{
\item{file_dir}{directory were file will be created, will create a dir for the game, and will try to put the plots there}
}
\value{

}
\description{
get_vornoi_animation
Will need to have gganimate installed
}
